requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
firebase-admin==6.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
python-jose==3.3.0
